Datasets/requirements.txt
numpy
pandas
pillow
fonttools
tqdm
opencv-python
